feat: Implement OAuth2 authentication and authorization across services
- Added OAuth2 client dependencies to api-gateway.
- Created ApiGatewayConfig and SecurityConfig for OAuth2 client setup in api-gateway.
- Implemented UserController to handle user info retrieval.
- Developed TokenRelayFilter to propagate OAuth2 tokens in api-gateway.
- Configured application.properties for OAuth2 client settings in api-gateway.
- Introduced spring-security-oauth2-authorization-server dependency in authentication-service.
- Created AuthorizationServerConfig for managing OAuth2 authorization server settings.
- Enhanced JwtConverterConfig and JwtDecoderConfig for JWT handling in authentication-service.
- Added JwtEncoderConfig for JWT encoding in authentication-service.
- Implemented TokenService for generating access and refresh tokens.
- Created OAuth2Controller for handling token requests and refreshes.
- Updated request DTO to include username for authentication requests.
- Added TokenResponse DTO for structured token responses.
- Implemented OAuth2AuditFilter for logging OAuth2 requests.
- Updated User model to use Long for ID.
- Modified AdminRepo and StaffRepo to use Long for ID.
- Configured ResourceServerConfig for patient-service and staff-service to secure endpoints.
- Updated application.properties for OAuth2 resource server settings in patient-service and staff-service.

+++ authentication-service/src/main/java/com/hcmutnightowls/authenticationservice/repository/AdminRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 @Repository
-public interface AdminRepo extends JpaRepository<Admin, Integer> {
+public interface AdminRepo extends JpaRepository<Admin, Long> {
     Optional<Admin> findBySubject(String subject);
 }
